zonesubscription_test.go
// File generated from our OpenAPI spec by Stainless.

package cloudflare_test

import (
	"context"
	"errors"
	"os"
	"testing"

	"github.com/cloudflare/cloudflare-sdk-go"
	"github.com/cloudflare/cloudflare-sdk-go/internal/testutil"
	"github.com/cloudflare/cloudflare-sdk-go/option"
)

func TestZoneSubscriptionZoneSubscriptionNewZoneSubscriptionWithOptionalParams(t *testing.T) {
	t.Skip("skipped: tests are disabled for the time being")
	baseURL := "http://localhost:4010"
	if envURL, ok := os.LookupEnv("TEST_API_BASE_URL"); ok {
		baseURL = envURL
	}
	if !testutil.CheckTestServer(t, baseURL) {
		return
	}
	client := cloudflare.NewClient(
		option.WithBaseURL(baseURL),
		option.WithAPIEmail("dev@cloudflare.com"),
		option.WithAPIKey("my-cloudflare-api-key"),
		option.WithAPIToken("my-cloudflare-api-token"),
		option.WithUserServiceKey("my-cloudflare-user-service-key"),
	)
	_, err := client.Zones.Subscriptions.ZoneSubscriptionNewZoneSubscription(
		context.TODO(),
		"506e3185e9c882d175a2d0cb0093d9f2",
		cloudflare.ZoneSubscriptionZoneSubscriptionNewZoneSubscriptionParams{
			App: cloudflare.F(cloudflare.ZoneSubscriptionZoneSubscriptionNewZoneSubscriptionParamsApp{
				InstallID: cloudflare.F("string"),
			}),
			ComponentValues: cloudflare.F([]cloudflare.ZoneSubscriptionZoneSubscriptionNewZoneSubscriptionParamsComponentValue{{
				Default: cloudflare.F(5.000000),
				Name:    cloudflare.F("page_rules"),
				Price:   cloudflare.F(5.000000),
				Value:   cloudflare.F(20.000000),
			}, {
				Default: cloudflare.F(5.000000),
				Name:    cloudflare.F("page_rules"),
				Price:   cloudflare.F(5.000000),
				Value:   cloudflare.F(20.000000),
			}, {
				Default: cloudflare.F(5.000000),
				Name:    cloudflare.F("page_rules"),
				Price:   cloudflare.F(5.000000),
				Value:   cloudflare.F(20.000000),
			}}),
			Frequency: cloudflare.F(cloudflare.ZoneSubscriptionZoneSubscriptionNewZoneSubscriptionParamsFrequencyMonthly),
			RatePlan: cloudflare.F(cloudflare.ZoneSubscriptionZoneSubscriptionNewZoneSubscriptionParamsRatePlan{
				Currency:          cloudflare.F("USD"),
				ExternallyManaged: cloudflare.F(false),
				ID:                cloudflare.F[any]("free"),
				IsContract:        cloudflare.F(false),
				PublicName:        cloudflare.F("Business Plan"),
				Scope:             cloudflare.F("zone"),
				Sets:              cloudflare.F([]string{"string", "string", "string"}),
			}),
			Zone: cloudflare.F(cloudflare.ZoneSubscriptionZoneSubscriptionNewZoneSubscriptionParamsZone{}),
		},
	)
	if err != nil {
		var apierr *cloudflare.Error
		if errors.As(err, &apierr) {
			t.Log(string(apierr.DumpRequest(true)))
		}
		t.Fatalf("err should be nil: %s", err.Error())
	}
}

func TestZoneSubscriptionZoneSubscriptionUpdateZoneSubscriptionWithOptionalParams(t *testing.T) {
	t.Skip("skipped: tests are disabled for the time being")
	baseURL := "http://localhost:4010"
	if envURL, ok := os.LookupEnv("TEST_API_BASE_URL"); ok {
		baseURL = envURL
	}
	if !testutil.CheckTestServer(t, baseURL) {
		return
	}
	client := cloudflare.NewClient(
		option.WithBaseURL(baseURL),
		option.WithAPIEmail("dev@cloudflare.com"),
		option.WithAPIKey("my-cloudflare-api-key"),
		option.WithAPIToken("my-cloudflare-api-token"),
		option.WithUserServiceKey("my-cloudflare-user-service-key"),
	)
	_, err := client.Zones.Subscriptions.ZoneSubscriptionUpdateZoneSubscription(
		context.TODO(),
		"506e3185e9c882d175a2d0cb0093d9f2",
		cloudflare.ZoneSubscriptionZoneSubscriptionUpdateZoneSubscriptionParams{
			App: cloudflare.F(cloudflare.ZoneSubscriptionZoneSubscriptionUpdateZoneSubscriptionParamsApp{
				InstallID: cloudflare.F("string"),
			}),
			ComponentValues: cloudflare.F([]cloudflare.ZoneSubscriptionZoneSubscriptionUpdateZoneSubscriptionParamsComponentValue{{
				Default: cloudflare.F(5.000000),
				Name:    cloudflare.F("page_rules"),
				Price:   cloudflare.F(5.000000),
				Value:   cloudflare.F(20.000000),
			}, {
				Default: cloudflare.F(5.000000),
				Name:    cloudflare.F("page_rules"),
				Price:   cloudflare.F(5.000000),
				Value:   cloudflare.F(20.000000),
			}, {
				Default: cloudflare.F(5.000000),
				Name:    cloudflare.F("page_rules"),
				Price:   cloudflare.F(5.000000),
				Value:   cloudflare.F(20.000000),
			}}),
			Frequency: cloudflare.F(cloudflare.ZoneSubscriptionZoneSubscriptionUpdateZoneSubscriptionParamsFrequencyMonthly),
			RatePlan: cloudflare.F(cloudflare.ZoneSubscriptionZoneSubscriptionUpdateZoneSubscriptionParamsRatePlan{
				Currency:          cloudflare.F("USD"),
				ExternallyManaged: cloudflare.F(false),
				ID:                cloudflare.F[any]("free"),
				IsContract:        cloudflare.F(false),
				PublicName:        cloudflare.F("Business Plan"),
				Scope:             cloudflare.F("zone"),
				Sets:              cloudflare.F([]string{"string", "string", "string"}),
			}),
			Zone: cloudflare.F(cloudflare.ZoneSubscriptionZoneSubscriptionUpdateZoneSubscriptionParamsZone{}),
		},
	)
	if err != nil {
		var apierr *cloudflare.Error
		if errors.As(err, &apierr) {
			t.Log(string(apierr.DumpRequest(true)))
		}
		t.Fatalf("err should be nil: %s", err.Error())
	}
}

func TestZoneSubscriptionZoneSubscriptionZoneSubscriptionDetails(t *testing.T) {
	t.Skip("skipped: tests are disabled for the time being")
	baseURL := "http://localhost:4010"
	if envURL, ok := os.LookupEnv("TEST_API_BASE_URL"); ok {
		baseURL = envURL
	}
	if !testutil.CheckTestServer(t, baseURL) {
		return
	}
	client := cloudflare.NewClient(
		option.WithBaseURL(baseURL),
		option.WithAPIEmail("dev@cloudflare.com"),
		option.WithAPIKey("my-cloudflare-api-key"),
		option.WithAPIToken("my-cloudflare-api-token"),
		option.WithUserServiceKey("my-cloudflare-user-service-key"),
	)
	_, err := client.Zones.Subscriptions.ZoneSubscriptionZoneSubscriptionDetails(context.TODO(), "506e3185e9c882d175a2d0cb0093d9f2")
	if err != nil {
		var apierr *cloudflare.Error
		if errors.As(err, &apierr) {
			t.Log(string(apierr.DumpRequest(true)))
		}
		t.Fatalf("err should be nil: %s", err.Error())
	}
}
